Move Segment to Segment::Single so classes aren't namespace Modules
MSP-10033

+++ spec/app/models/metasploit_data_models/ip_address/v4/segment/single_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe MetasploitDataModels::IPAddress::V4::Segment do
+describe MetasploitDataModels::IPAddress::V4::Segment::Single do
   subject(:number) {
     described_class.new(
         value: formatted_value
